Users/schemas.py
from typing import Annotated
from annotated_types import MinLen, MaxLen
from pydantic import BaseModel, EmailStr


class CreateUser(BaseModel):
    username: Annotated[str, MaxLen(20), MinLen(3)]
    email: EmailStr
