pyproject.toml
[project]
name = "shop"
version = "0.1.0"
description = ""
authors = [
    {name = "maksk",email = "wgagaqwt1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

